smart_contracts/config.py
import dataclasses
from collections.abc import Callable

from algokit_utils import Account, ApplicationSpecification
from algosdk.v2client.algod import AlgodClient
from algosdk.v2client.indexer import IndexerClient
from beaker import Application

from smart_contracts.insurance.contract import app as insurance_app
from smart_contracts.insurance.deploy_config import deploy as insurance_deploy


@dataclasses.dataclass
class SmartContract:
    app: Application
    deploy: Callable[
        [AlgodClient, IndexerClient, ApplicationSpecification, Account], None
    ] | None = None


# define contracts to build and/or deploy
contracts = [SmartContract(app=insurance_app, deploy=insurance_deploy)]
